Banking_System/requirements.txt
asgiref==3.6.0
boto3==1.26.101
botocore==1.29.101
Django==4.1.7
django-cors-headers==3.14.0
django-storages==1.13.2
djangorestframework==3.14.0
gunicorn==20.1.0
jmespath==1.0.1
PyJWT==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.2
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.15
